requirements.txt
# Web framework
fastapi>=0.110
uvicorn>=0.27

# HTTP client
requests>=2.31

# LangGraph / LangChain core
langgraph
langchain-core

# Async / background processing
celery>=5.3
redis>=5.0

# Configuration / settings
pydantic>=2.5
pydantic-settings>=2.1

alembic
python-jose[cryptography]